+++ lang/csharp/Continental.eCAL.Core/CMakeLists.txt
@@ -28,23 +28,21 @@ set(ecal_clr_src
     ecal_clr_common.h
     ecal_clr_core.cpp
     ecal_clr_core.h
-    ecal_clr_datatype.h
     ecal_clr_log.cpp
     ecal_clr_log.h
     ecal_clr_monitoring.cpp
     ecal_clr_monitoring.h
     ecal_clr_publisher.cpp
     ecal_clr_publisher.h
-    ecal_clr_receivecallbackdata.h
     ecal_clr_server.cpp
     ecal_clr_server.h
-    ecal_clr_servicetypes.h
     ecal_clr_subscriber.cpp
     ecal_clr_subscriber.h
-    ecal_clr_subscribercallback.h
     ecal_clr_time.cpp
     ecal_clr_time.h
-    ecal_clr_topicid.h
+    ecal_clr_types.h
+    ecal_clr_types_pubsub.h
+    ecal_clr_types_service.h
 ) 
 
 add_library(${PROJECT_NAME} SHARED ${ecal_clr_src})
