+++ lang/csharp/Continental.eCAL.Core/ecal_clr_subscriber.cpp
@@ -20,10 +20,8 @@
 #include "ecal_clr_subscriber.h"
 
 #include "ecal_clr_common.h"
-#include "ecal_clr_datatype.h"
-#include "ecal_clr_topicid.h"
-#include "ecal_clr_subscribercallback.h"
-#include "ecal_clr_receivecallbackdata.h"
+#include "ecal_clr_types.h"
+#include "ecal_clr_types_pubsub.h"
 
 #include <sstream>
 #include <vcclr.h>
