+++ lang/csharp/Continental.eCAL.Core/ecal_clr_types.h
